todolist/requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.2
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
drf-spectacular==0.24.2
exceptiongroup==1.0.4
factory-boy==3.2.1
Faker==15.3.2
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
jsonschema==4.17.0
oauthlib==3.2.2
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.4
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.5
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
